add function exercises day 3

+++ Day_3/Loops_Exercises/Challenge_10.c
@@ -15,7 +15,7 @@ int main()
 	int i = 0; 
 
 
-	srand((unsigned int)(time(NULL))); 
+	srand(time(NULL)); 
 
 	char generateur[62] = "0123456789abcdefghijklmnoqprstuvwyzxABCDEFGHIJKLMNOQPRSTUYWVZX";
